requirements.txt
tensorflow==2.12.0
numpy==1.23.5
deepface==0.0.75
opencv-python==4.8.0.76
pillow==10.0.0
flask==2.3.3
python-dotenv==1.0.0
moviepy==1.0.3
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0 